--- a/AirLib/include/common/AirSimSettings.hpp
+++ b/AirLib/include/common/AirSimSettings.hpp
@@ -249,6 +249,11 @@ namespace airlib
 
         struct LidarSetting : SensorSetting
         {
+            enum class DataFrame
+            {
+                VehicleInertialFrame,
+                SensorLocalFrame
+            };
         };
 
         struct VehicleSetting