--- a/Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealLidarSensor.h
+++ b/Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealLidarSensor.h
@@ -31,7 +31,7 @@ private:
     void createLasers();
     bool shootLaser(const msr::airlib::Pose& lidar_pose, const msr::airlib::Pose& vehicle_pose,
                     const float horizontal_angle, const float vertical_angle,
-                    const msr::airlib::LidarSimpleParams params, Vector3r& point, int& segmentationID);
+                    const msr::airlib::LidarSimpleParams& params, Vector3r& point, int& segmentationID);
 
 private:
     AActor* actor_;