<!-- Thank you for submitting a pull request! -->
<!-- ⚠️⚠️ Do Not Delete This! pull_request_template ⚠️⚠️ -->
<!-- Please read our contribution guidelines: https://microsoft.github.io/AirSim/CONTRIBUTING/ -->

## About
<!-- Describe what your PR is about. -->
Lidar raycast make the calculation using line trace on single thread.
This PR take advantage of `ParallelFor` to improve the performance.

## How Has This Been Tested?
<!-- Please, describe how you have tested your changes to help us incorporate them. -->
<details>
<summary>Lidar config tested</summary>

```

{
  "SettingsVersion": 1.2,
  "SimMode": "Car",
  "EnableRpc": true,
  "ViewMode": "Fpv",
  "LogMessagesVisible": true,
  "EngineSound": false,

  "Vehicles": {
    "PhysXCar": {
      "VehicleType": "PhysXCar",
      "DefaultVehicleState": "",
      "AutoCreate": true,
      "PawnPath": "",
      "EnableCollisionPassthrogh": false,
      "EnableCollisions": true,
      "RC": {
        "RemoteControlID": 0
      },
      "Sensors": {
        "LidarCustom": {
            "SensorType": 6,
            "Enabled" : true,
            "NumberOfChannels": 16,
            "PointsPerSecond": 32000,
            "Range": 40.0,
            "X": 2.0, "Y": 0, "Z": -0.7657162547111511,
            "Pitch": 0.0, "Roll": 0.0, "Yaw": 0.0,
            "HorizontalFOVStart": -180.0,
            "HorizontalFOVEnd": 180.0,
            "VerticalFOVUpper": 45.0,
            "VerticalFOVLower": -10.0,     
            "DrawDebugPoints": true
        }
      }
    }
  }
}

```
</details>

**Benchmark:**
<details>
<summary>Before:</summary>

```
LogTemp: Warning: code executed in 0.023679 seconds.
LogTemp: Warning: code executed in 0.007037 seconds.
LogTemp: Warning: code executed in 0.007560 seconds.
LogTemp: Warning: code executed in 0.007134 seconds.
LogTemp: Warning: code executed in 0.008667 seconds.
LogTemp: Warning: code executed in 0.007389 seconds.
LogTemp: Warning: code executed in 0.007219 seconds.
LogTemp: Warning: code executed in 0.008543 seconds.
LogTemp: Warning: code executed in 0.007057 seconds.
LogTemp: Warning: code executed in 0.009526 seconds.
LogTemp: Warning: code executed in 0.008571 seconds.
LogTemp: Warning: code executed in 0.007375 seconds.
LogTemp: Warning: code executed in 0.007126 seconds.
LogTemp: Warning: code executed in 0.008792 seconds.
LogTemp: Warning: code executed in 0.006901 seconds.
LogTemp: Warning: code executed in 0.007225 seconds.
LogTemp: Warning: code executed in 0.008875 seconds.
LogTemp: Warning: code executed in 0.007040 seconds.
LogTemp: Warning: code executed in 0.007565 seconds.
LogTemp: Warning: code executed in 0.008645 seconds.
LogTemp: Warning: code executed in 0.008770 seconds.
LogTemp: Warning: code executed in 0.007121 seconds.
LogTemp: Warning: code executed in 0.008888 seconds.
LogTemp: Warning: code executed in 0.009491 seconds.
LogTemp: Warning: code executed in 0.008847 seconds.
LogTemp: Warning: code executed in 0.008941 seconds.
LogTemp: Warning: code executed in 0.007436 seconds.
LogTemp: Warning: code executed in 0.009542 seconds.
LogTemp: Warning: code executed in 0.009030 seconds.
LogTemp: Warning: code executed in 0.008122 seconds.
LogTemp: Warning: code executed in 0.009491 seconds.
LogTemp: Warning: code executed in 0.008235 seconds.
LogTemp: Warning: code executed in 0.010218 seconds.
LogTemp: Warning: code executed in 0.008488 seconds.
LogTemp: Warning: code executed in 0.010495 seconds.
LogTemp: Warning: code executed in 0.009934 seconds.
LogTemp: Warning: code executed in 0.014051 seconds.

```
</details>

<details>
<summary>After:</summary>

```
LogTemp: Warning: code executed in 0.002308 seconds.
LogTemp: Warning: code executed in 0.001656 seconds.
LogTemp: Warning: code executed in 0.001599 seconds.
LogTemp: Warning: code executed in 0.001483 seconds.
LogTemp: Warning: code executed in 0.001626 seconds.
LogTemp: Warning: code executed in 0.001331 seconds.
LogTemp: Warning: code executed in 0.001581 seconds.
LogTemp: Warning: code executed in 0.001809 seconds.
LogTemp: Warning: code executed in 0.001552 seconds.
LogTemp: Warning: code executed in 0.001292 seconds.
LogTemp: Warning: code executed in 0.001149 seconds.
LogTemp: Warning: code executed in 0.002029 seconds.
LogTemp: Warning: code executed in 0.001522 seconds.
LogTemp: Warning: code executed in 0.001314 seconds.
LogTemp: Warning: code executed in 0.001266 seconds.
LogTemp: Warning: code executed in 0.001196 seconds.
LogTemp: Warning: code executed in 0.002130 seconds.
LogTemp: Warning: code executed in 0.001464 seconds.
LogTemp: Warning: code executed in 0.001435 seconds.
LogTemp: Warning: code executed in 0.001353 seconds.
LogTemp: Warning: code executed in 0.001157 seconds.
LogTemp: Warning: code executed in 0.002004 seconds.
LogTemp: Warning: code executed in 0.001739 seconds.
LogTemp: Warning: code executed in 0.001274 seconds.
LogTemp: Warning: code executed in 0.001429 seconds.
LogTemp: Warning: code executed in 0.002372 seconds.
LogTemp: Warning: code executed in 0.002305 seconds.
LogTemp: Warning: code executed in 0.001499 seconds.
LogTemp: Warning: code executed in 0.001063 seconds.
LogTemp: Warning: code executed in 0.001904 seconds.
LogTemp: Warning: code executed in 0.001390 seconds.
LogTemp: Warning: code executed in 0.001310 seconds.
LogTemp: Warning: code executed in 0.001293 seconds.
LogTemp: Warning: code executed in 0.001710 seconds.
LogTemp: Warning: code executed in 0.001767 seconds.
LogTemp: Warning: code executed in 0.001296 seconds.
LogTemp: Warning: code executed in 0.001814 seconds.
LogTemp: Warning: code executed in 0.001432 seconds.
LogTemp: Warning: code executed in 0.001237 seconds.
LogTemp: Warning: code executed in 0.001295 seconds.
LogTemp: Warning: code executed in 0.001241 seconds.
LogTemp: Warning: code executed in 0.001401 seconds.
LogTemp: Warning: code executed in 0.001784 seconds.
LogTemp: Warning: code executed in 0.001396 seconds.

```
</details>

You can see the execution time is about 8-9 times faster, which obviously depends on your CPU.