@rajat2004 2022-05-08
Some more things which might also help in improving performance -

https://github.com/microsoft/AirSim/blob/19ac0f265e0b81031c173ded3444be052ddaab8e/Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealLidarSensor.cpp#L139-L141

This is just getting the first component with id != -1, instead of continuing to iterate, we can break early -

```
for (auto* comp : meshCompoenents)
{
    segmentationID = comp->CustomDepthStencilValue;
    if (segmentationID != -1)
        break;
}
```

https://github.com/microsoft/AirSim/blob/19ac0f265e0b81031c173ded3444be052ddaab8e/Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealLidarSensor.cpp#L105-L107

Here `laser` is unused in the method body, should be removed. Also, `const msr::airlib::LidarSimpleParams params` is getting copied every time, that's also a performance hit

https://github.com/microsoft/AirSim/blob/19ac0f265e0b81031c173ded3444be052ddaab8e/Unreal/Plugins/AirSim/Source/UnrealSensors/UnrealLidarSensor.cpp#L49

This could be made a const ref