@rajat2004 2022-05-08
Another thing which I just wanted to note, from my (limited) experience with parallel programming, it was often better to have lesser number of parallel tasks each doing more work, rather than large number of small tasks. This helped in reducing core contention and improving cache locality as well.
Rather than doing ParallelFor for `total_points_to_scan`, it might be worth doing some batching, like `total_points_to_scan / batch_size`, and it would work on `batch_size` number of points. It would complicate the code a bit though, just something which might be interesting, doesn't need to be done in this PR. If trying this out, making it a configurable param would help in finding appropriate sizes which work well