@xxEoD2242 2022-08-04
> Since this project is going to be archived, could this pull request be merged in the [Colosseum fork?](https://github.com/CodexLabsLLC/Colosseum)

Hey @wouter-heerwegh,

I'm happy to merge this in. Give me a minute to figure out how to pull this PR into the forked repo and I will.