@jonyMarino 2025-08-11
[Project AirSim](https://github.com/iamaisim/ProjectAirSim/) now features a GPU-accelerated LiDAR sensor with real-world sensor models.
This means you can simulate LiDAR data at high speed while maintaining realistic physics and sensor behavior, closely matching actual hardware performance.

With GPU acceleration, you can run high-resolution scans in real time, making it ideal for testing:

Autonomous navigation and SLAM

Obstacle detection and avoidance

Multi-vehicle sensor fusion

By modeling real LiDAR devices, Project AirSim enables more accurate algorithm development and faster deployment from simulation to real-world robotics.